+++ src/path/functions/index.ts
@@ -5,3 +5,5 @@ export { Search } from "./search";
 export { Value } from "./value";
 export { FunctionExpressionType } from "./function";
 export type { FilterFunction } from "./function";
+export type { MatchFilterFunctionOptions } from "./match";
+export type { SearchFilterFunctionOptions } from "./search";