+++ src/path/functions/match.ts
@@ -1,5 +1,9 @@
+import { isString } from "../../types";
+import { IRegexpError } from "../errors";
 import { LRUCache } from "../lru_cache";
 import { FilterFunction, FunctionExpressionType } from "./function";
+import { mapRegexp } from "./pattern";
+import { check } from "iregexp-check";
 
 export type MatchFilterFunctionOptions = {
   /**
@@ -8,11 +12,21 @@ export type MatchFilterFunctionOptions = {
   cacheSize?: number;
 
   /**
-   * If _true_, throw errors from regex construction and matching.
-   * The standard and default behavior is to ignore these errors
-   * and return _false_.
+   * If _true_, throw errors from regex checking, construction and matching.
+   * The standard and default behavior is to ignore these errors and return
+   * _false_.
    */
   throwErrors?: boolean;
+
+  /**
+   * If _true_, check that regexp patterns are valid according to I-Regexp.
+   * The standard and default behavior is to silently return _false_ if a
+   * pattern is invalid.
+   *
+   * If `iRegexpCheck` is _true_ and `throwErrors` is _true_, an `IRegexpError`
+   * will be thrown.
+   */
+  iRegexpCheck?: boolean;
 };
 
 export class Match implements FilterFunction {
@@ -25,14 +39,17 @@ export class Match implements FilterFunction {
 
   readonly cacheSize: number;
   readonly throwErrors: boolean;
+  readonly iRegexpCheck: boolean;
   #cache: LRUCache<string, RegExp>;
 
   constructor(readonly options: MatchFilterFunctionOptions = {}) {
     this.cacheSize = options.cacheSize ?? 10;
     this.throwErrors = options.throwErrors ?? false;
+    this.iRegexpCheck = options.iRegexpCheck ?? true;
     this.#cache = new LRUCache(this.cacheSize);
   }
 
+  // eslint-disable-next-line sonarjs/cognitive-complexity
   public call(s: string, pattern: string): boolean {
     if (this.cacheSize > 0) {
       const re = this.#cache.get(pattern);
@@ -44,6 +61,24 @@ export class Match implements FilterFunction {
           return false;
         }
       }
+    }
+
+    if (!isString(pattern)) {
+      if (this.throwErrors) {
+        throw new IRegexpError(
+          `match() expected a string pattern, found ${pattern}`,
+        );
+      }
+      return false;
+    }
+
+    if (this.iRegexpCheck && !check(pattern)) {
+      if (this.throwErrors) {
+        throw new IRegexpError(
+          `pattern ${pattern} is not a valid I-Regexp pattern`,
+        );
+      }
+      return false;
     }
 
     try {
@@ -58,56 +93,11 @@ export class Match implements FilterFunction {
 
   protected fullMatch(pattern: string): string {
     const parts: string[] = [];
-    let nonCaptureGroup = false;
-
-    if (!pattern.startsWith("^") && !pattern.startsWith("^(")) {
-      nonCaptureGroup = true;
-      parts.push("^(?:");
-    }
-    parts.push(this.mapRegexp(pattern));
-
-    if (nonCaptureGroup && !pattern.endsWith("$") && !pattern.endsWith(")$")) {
-      parts.push(")$");
-    }
-
-    return parts.join("");
-  }
-
-  // See https://datatracker.ietf.org/doc/html/rfc9485#name-ecmascript-regexps
-  protected mapRegexp(pattern: string): string {
-    let escaped = false;
-    let charClass = false;
-    const parts: string[] = [];
-    for (const ch of pattern) {
-      switch (ch) {
-        case ".":
-          if (!escaped && !charClass) {
-            parts.push("(?:(?![\r\n])\\P{Cs}|\\p{Cs}\\p{Cs})");
-          } else {
-            parts.push(ch);
-            escaped = false;
-          }
-          break;
-        case "\\":
-          escaped = true;
-          parts.push(ch);
-          break;
-        case "[":
-          charClass = true;
-          escaped = false;
-          parts.push(ch);
-          break;
-        case "]":
-          charClass = false;
-          escaped = false;
-          parts.push(ch);
-          break;
-        default:
-          escaped = false;
-          parts.push(ch);
-          break;
-      }
-    }
+    const explicitCaret = pattern.startsWith("^");
+    const explicitDollar = pattern.endsWith("$");
+    if (!explicitCaret && !explicitDollar) parts.push("^(?:");
+    parts.push(mapRegexp(pattern));
+    if (!explicitCaret && !explicitDollar) parts.push(")$");
     return parts.join("");
   }
 }