+++ src/path/functions/search.ts
@@ -1,5 +1,9 @@
+import { check } from "iregexp-check";
 import { LRUCache } from "../lru_cache";
 import { FilterFunction, FunctionExpressionType } from "./function";
+import { mapRegexp } from "./pattern";
+import { IRegexpError } from "../errors";
+import { isString } from "../../types";
 
 export type SearchFilterFunctionOptions = {
   /**
@@ -14,6 +18,16 @@ export type SearchFilterFunctionOptions = {
    * and return _false_.
    */
   throwErrors?: boolean;
+
+  /**
+   * If _true_, check that regexp patterns are valid according to I-Regexp.
+   * The standard and default behavior is to silently return _false_ if a
+   * pattern is invalid.
+   *
+   * If `iRegexpCheck` is _true_ and `throwErrors` is _true_, an `IRegexpError`
+   * will be thrown.
+   */
+  iRegexpCheck?: boolean;
 };
 
 export class Search implements FilterFunction {
@@ -26,14 +40,17 @@ export class Search implements FilterFunction {
 
   readonly cacheSize: number;
   readonly throwErrors: boolean;
+  readonly iRegexpCheck: boolean;
   #cache: LRUCache<string, RegExp>;
 
   constructor(readonly options: SearchFilterFunctionOptions = {}) {
     this.cacheSize = options.cacheSize ?? 10;
     this.throwErrors = options.throwErrors ?? false;
+    this.iRegexpCheck = options.iRegexpCheck ?? true;
     this.#cache = new LRUCache(this.cacheSize);
   }
 
+  // eslint-disable-next-line sonarjs/cognitive-complexity
   public call(s: string, pattern: string): boolean {
     if (this.cacheSize > 0) {
       const re = this.#cache.get(pattern);
@@ -47,51 +64,31 @@ export class Search implements FilterFunction {
       }
     }
 
+    if (!isString(pattern)) {
+      if (this.throwErrors) {
+        throw new IRegexpError(
+          `match() expected a string pattern, found ${pattern}`,
+        );
+      }
+      return false;
+    }
+
+    if (this.iRegexpCheck && !check(pattern)) {
+      if (this.throwErrors) {
+        throw new IRegexpError(
+          `pattern ${pattern} is not a valid I-Regexp pattern`,
+        );
+      }
+      return false;
+    }
+
     try {
-      const re = new RegExp(this.mapRegexp(pattern), "u");
+      const re = new RegExp(mapRegexp(pattern), "u");
       if (this.cacheSize > 0) this.#cache.set(pattern, re);
       return !!s.match(re);
     } catch (error) {
       if (this.throwErrors) throw error;
       return false;
     }
-  }
-
-  // See https://datatracker.ietf.org/doc/html/rfc9485#name-ecmascript-regexps
-  protected mapRegexp(pattern: string): string {
-    let escaped = false;
-    let charClass = false;
-    const parts: string[] = [];
-    for (const ch of pattern) {
-      switch (ch) {
-        case ".":
-          if (!escaped && !charClass) {
-            parts.push("(?:(?![\r\n])\\P{Cs}|\\p{Cs}\\p{Cs})");
-          } else {
-            parts.push(ch);
-            escaped = false;
-          }
-          break;
-        case "\\":
-          escaped = true;
-          parts.push(ch);
-          break;
-        case "[":
-          charClass = true;
-          escaped = false;
-          parts.push(ch);
-          break;
-        case "]":
-          charClass = false;
-          escaped = false;
-          parts.push(ch);
-          break;
-        default:
-          escaped = false;
-          parts.push(ch);
-          break;
-      }
-    }
-    return parts.join("");
   }
 }