+++ tests/path/regex_filters.test.ts
@@ -1,4 +1,5 @@
 import { JSONPathEnvironment } from "../../src/path";
+import { IRegexpError } from "../../src/path/errors";
 import { Match, Search } from "../../src/path/functions";
 
 describe("match filter", () => {
@@ -23,9 +24,59 @@ describe("match filter", () => {
       new Match({ cacheSize: 0, throwErrors: true }),
     );
     expect(() => env.query("$[?match(@.a, 'a.*(')]", [{ a: "ab" }])).toThrow(
-      SyntaxError,
+      IRegexpError,
     );
   });
+  test("don't replace dot in character group", () => {
+    const env = new JSONPathEnvironment();
+    const query = "$[?match(@, 'ab[.c]d')]";
+    const data = ["abcd", "ab.d", "abxd"];
+    const rv = env.query(query, data);
+    expect(rv.values()).toStrictEqual(["abcd", "ab.d"]);
+  });
+  test("don't replace escaped dots", () => {
+    const env = new JSONPathEnvironment();
+    const query = "$[?match(@, 'ab\\\\.d')]";
+    const data = ["abcd", "ab.d", "abxd"];
+    const rv = env.query(query, data);
+    expect(rv.values()).toStrictEqual(["ab.d"]);
+  });
+  test("handle escaped right square bracket in character group", () => {
+    const env = new JSONPathEnvironment();
+    const query = "$[?match(@, 'ab[\\\\].c]d')]";
+    const data = ["abcd", "ab.d", "abxd"];
+    const rv = env.query(query, data);
+    expect(rv.values()).toStrictEqual(["abcd", "ab.d"]);
+  });
+  test("explicit start caret", () => {
+    const env = new JSONPathEnvironment();
+    const query = "$[?match(@, '^ab.*')]";
+    const data = ["abcd", "ab.d", "axc"];
+    const rv = env.query(query, data);
+    expect(rv.values()).toStrictEqual(["abcd", "ab.d"]);
+  });
+  test("explicit end dollar", () => {
+    const env = new JSONPathEnvironment();
+    const query = "$[?match(@, '.bc$')]";
+    const data = ["abcd", "abc", "axc"];
+    const rv = env.query(query, data);
+    expect(rv.values()).toStrictEqual(["abc"]);
+  });
+  // test("handle escaped left square bracket", () => {
+  //   const env = new JSONPathEnvironment();
+  //   const query = "$[?match(@, 'ab\\\\[.d')]";
+  //   const data = ["abcd", "ab.d", "ab[d"];
+  //   const rv = env.query(query, data);
+  //   expect(rv.values()).toStrictEqual(["ab[d"]);
+  // });
+
+  // test("handle escaped backslash before dot", () => {
+  //   const env = new JSONPathEnvironment();
+  //   const query = "$[?match(@, 'ab\\\\\\\\.d')]";
+  //   const data = ["abcd", "ab.d", "ab\\d"];
+  //   const rv = env.query(query, data);
+  //   expect(rv.values()).toStrictEqual(["ab\\d"]);
+  // });
 });
 
 describe("search filter", () => {
@@ -51,6 +102,6 @@ describe("search filter", () => {
     );
     expect(() =>
       env.query("$[?search(@.a, 'a.*(')]", [{ a: "the end is ab" }]),
-    ).toThrow(SyntaxError);
+    ).toThrow(IRegexpError);
   });
 });